backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
ultralytics==8.0.196
supervision==0.16.0
torch==2.1.1
torchvision==0.16.1
openai==1.3.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1 